aldor.language: Add a first pass at an Aldor type model.
As it stands, this will read in .asy files and be able to answer questions
about type exports and parameters.

Design is pretty similar to the aldor compiler - differences are mostly
because java is a somewhat different language.

Anyway, idea of this (as well as being handy in eclipse), is that it can
serve as a model re-implementation in other languages (eg. Aldor).